Add local testing data and routes.
