+++ .env.development
@@ -1,0 +1,4 @@
+BASE_URL="/plot-navigator"
+REPO_URLS={"debug": ""}
+PAGE_LINKS="Metrics,Bokeh"
+ENABLE_TEST_IMAGES="true"
